+++ tests/test_getcwd.py
@@ -1,14 +1,12 @@
-'''test sftpretty.getcwd
-until you issue a .chdir/cwd command paramiko returns None for .getcwd,
-unless you have set default_path in the Connection args'''
+'''test sftpretty.getcwd'''
 
 from common import conn, VFS
+from pathlib import Path
 from sftpretty import Connection
 
 
 def test_getcwd_none(sftpserver):
-    '''test .getcwd as the first operation - need pristine connection
-    and no default_path arg'''
+    '''test .getcwd with no default_path arg'''
     with sftpserver.serve_content(VFS):
         cnn = conn(sftpserver)
         cnn['default_path'] = None
@@ -25,9 +23,10 @@ def test_getcwd_default_path(sftpserver):
 
 def test_getcwd_after_chdir(sftpserver):
     '''test getcwd after a chdir operation'''
+    pubpath = Path('/home/test').joinpath('pub/foo1')
     with sftpserver.serve_content(VFS):
         cnn = conn(sftpserver)
         cnn['default_path'] = None
         with Connection(**cnn) as sftp:
-            sftp.chdir('/home/test/pub/foo1')
-            assert sftp.getcwd() == '/home/test/pub/foo1'
+            sftp.chdir(pubpath.as_posix())
+            assert sftp.getcwd() == pubpath.as_posix()