+++ tests/test_issue_65.py
@@ -2,18 +2,20 @@
 location'''
 
 from common import conn, VFS
+from pathlib import Path
 from sftpretty import Connection
 
 
 def test_issue_65(sftpserver):
     '''using the .cd() context manager prior to setting a directory
     via chdir causes an error'''
+    pubpath = Path('/home/test').joinpath('pub')
     with sftpserver.serve_content(VFS):
         cnn = conn(sftpserver)
-        cnn['default_path'] = None  # don't call .chdir by setting default_path
+        cnn['default_path'] = None
         with Connection(**cnn) as sftp:
             assert sftp.getcwd() is None
-            with sftp.cd('/home/test/pub'):
+            with sftp.cd(pubpath.as_posix()):
                 pass
 
             assert sftp.getcwd() == '/'