+++ tests/test_remotetree.py
@@ -11,25 +11,31 @@ def test_remotetree(sftpserver):
     with sftpserver.serve_content(VFS):
         with Connection(**conn(sftpserver)) as sftp:
             cwd = sftp.pwd
-            directories = {}
             localpath = Path(mkdtemp()).as_posix()
+            tree = {}
 
-            sftp.remotetree(directories, cwd, localpath)
+            sftp.remotetree(tree, cwd, localpath)
 
-            dkeys = ['/home/test',
-                     '/home/test/pub',
-                     '/home/test/pub/foo2']
+            remote = {
+                '/home/test': [
+                    ('/home/test/pub', f'{localpath}/pub')
+                ],
+                '/home/test/pub': [
+                    ('/home/test/pub/foo1',
+                     f'{localpath}/pub/foo1'),
+                    ('/home/test/pub/foo2',
+                     f'{localpath}/pub/foo2')
+                ],
+                '/home/test/pub/foo2': [
+                    ('/home/test/pub/foo2/bar1',
+                     f'{localpath}/pub/foo2/bar1')
+                ]
+            }
 
-            dvalues = [[('/home/test/pub', f'{localpath}/home/test/pub')],
-                       [('/home/test/pub/foo1',
-                         f'{localpath}/home/test/pub/foo1'),
-                        ('/home/test/pub/foo2',
-                         f'{localpath}/home/test/pub/foo2')],
-                       [('/home/test/pub/foo2/bar1',
-                         f'{localpath}/home/test/pub/foo2/bar1')]]
-
-            assert list(directories.keys()) == dkeys
-            assert list(directories.values()) == dvalues
+            for branch in sorted(tree.keys()):
+                assert set(remote[branch]) == set(tree[branch])
+                del tree[branch]
+            assert tree == {}
 
 
 def test_remotetree_no_recurse(sftpserver):
@@ -37,13 +43,18 @@ def test_remotetree_no_recurse(sftpserver):
     with sftpserver.serve_content(VFS):
         with Connection(**conn(sftpserver)) as sftp:
             cwd = sftp.pwd
-            directories = {}
             localpath = Path(mkdtemp()).as_posix()
+            tree = {}
 
-            sftp.remotetree(directories, cwd, localpath, recurse=False)
+            sftp.remotetree(tree, cwd, localpath, recurse=False)
 
-            dkeys = ['/home/test']
-            dvalues = [[('/home/test/pub', f'{localpath}/home/test/pub')]]
+            remote = {
+                '/home/test': [
+                    ('/home/test/pub', f'{localpath}/pub')
+                ]
+            }
 
-            assert list(directories.keys()) == dkeys
-            assert list(directories.values()) == dvalues
+            for branch in sorted(tree.keys()):
+                assert set(remote[branch]) == set(tree[branch])
+                del tree[branch]
+            assert tree == {}