+++ tests/test_truncate.py
@@ -1,10 +1,9 @@
 '''test sftpretty.listdir'''
 
-from common import SKIP_IF_CI, STARS8192
+from common import STARS8192
 from io import BytesIO
 
 
-@SKIP_IF_CI
 def test_truncate_smaller(lsftp):
     '''test truncate, make file smaller'''
     flo = BytesIO(bytes(STARS8192, 'UTF-8'))
@@ -19,7 +18,6 @@ def test_truncate_smaller(lsftp):
     lsftp.remove(rname)
 
 
-@SKIP_IF_CI
 def test_truncate_larger(lsftp):
     '''test truncate, make file larger'''
     flo = BytesIO(bytes(STARS8192, 'UTF-8'))
@@ -34,7 +32,6 @@ def test_truncate_larger(lsftp):
     lsftp.remove(rname)
 
 
-@SKIP_IF_CI
 def test_truncate_same(lsftp):
     '''test truncate, make file same size'''
     flo = BytesIO(bytes(STARS8192, 'UTF-8'))
@@ -52,6 +49,6 @@ def test_truncate_same(lsftp):
 # TODO
 # def test_truncate_ro(psftp):
 #     '''test truncate, against read-only server'''
-#     rname = '/home/test/readme.txt'
+#     rname = Path.home().joinpath('readme.txt').as_posix()
 #     with pytest.raises(IOError):
 #         _ = psftp.truncate(rname, 8192)